requirements.txt
torchnet==0.0.4
tqdm==4.36.1
colorlog==4.1.0
torch==1.5.0
matplotlib==3.1.1
numpy==1.18.4
prettytable==0.7.2
torchvision==0.6.0
Pillow==7.2.0
PyYAML==5.3.1
tensorboardX==2.1
